refactor(companion): split the settings screen into per-card widgets
settings_screen.dart was 3,434 lines — twice the next-largest Dart file
— with one State object owning voice, appearance, mesh, and overview
concerns alongside the daemon-config cards and the two layouts. The
per-card builders were already independent; they are now widgets in
lib/src/ui/settings/:

- voice_card.dart — VoiceCard: the voice list, the assistant-role probe
  (lifecycle observer), preview playback (stopped on dispose), the
  picker sheet and the locale/quality label helpers.
- appearance_card.dart — AppearanceCard: theme mode, accent swatches
  and the custom picker, text size, haptics/notification switches; owns
  the resolved wallpaper accent.
- mesh_card.dart — MeshCard: sharing/reporting/device-control switches,
  the Android privilege ladder and adb-root flow, desktop start-at-login,
  and the device list.
- overview_cards.dart — StatusCard, DiagnosticsCard, AboutCard (loads its
  own version), ConnectionCard.
- settings_widgets.dart — the shared frame and rows (SettingsSection,
  settingsSwitchRow/TextRow/IntervalRow, healthRow, infoRow, statTile,
  ControlButton, ModelRow, RailTile, SettingsChapter, SettingsSkeleton,
  fmtUptime/fmtAge, SettingsHealth).

settings_screen.dart keeps what is genuinely the screen's: the config
snapshot and optimistic-update state, load/apply, the controls, the
General/Background agents/Extensions/Controls cards, and both layouts.
It is 790 lines. Code moved verbatim apart from widget.state → state in
stateless cards, the nullable cfg promoted through a local, and one
build-context guard that now checks context.mounted because the
closure's context is the build parameter rather than State.context.

One behavioural nuance: each card's side state (voice list, wallpaper
accent, privilege ladder) is now initialised when the card mounts, so
on the wide layout switching chapters re-probes them and a running
voice preview stops with its chapter. The phone column is unchanged.

flutter analyze: no issues. flutter test: 246 passed.
